requirements.txt
django==5.2.2
flake8==7.3.0
pytest-django==4.11.1
pytest-cov==6.3.0
whitenoise==6.9.0
